+++ components/AppForm.vue
@@ -1,18 +1,9 @@
 <script lang="ts" setup>
-import { parseISO } from 'date-fns'
-
 const formStore = useFormStore()
 const onClickTwilogSearchButton = () => {
   const url = createTwilogKeywordSearchURL({
     user: formStore.user,
     keyword: formStore.keyword
-  })
-  window.open(url)
-}
-const onClickTwisaveDateButton = () => {
-  const url = createTwisaveDateURL({
-    user: formStore.user,
-    date: parseISO(formStore.date)
   })
   window.open(url)
 }
@@ -38,11 +29,7 @@ const onClickTwisaveSearchButton = () => {
         Twilog（検索）
       </el-button>
     </el-form-item>
-    <el-form-item>
-      <el-button @click="onClickTwisaveDateButton">
-        ツイセーブ（日付）
-      </el-button>
-    </el-form-item>
+    <AppFormTwisaveDate />
     <el-form-item>
       <el-button @click="onClickTwisaveSearchButton">
         ツイセーブ（検索）
